requirements.txt
psycopg2-binary==2.9
Faker~=13.4.0
django==3.2
flake8==4.0
python-dotenv==0.20
django-split-settings==1.1
django-debug-toolbar==3.2.4
pytest

pydantic==1.9.0